Refactor whois record tests to be separate by type of record
Also, remove one nil check from WhoisServer class
