+++ lib/whois_server.rb
@@ -49,15 +49,12 @@ module WhoisServer
     name = SimpleIDN.to_unicode(name)
     whois_record = WhoisRecord.find_by(name: name)
 
-    if whois_record.nil?
-      logger.info "#{ip}: requested: #{data} [searched by: #{name}; No record found]"
-      send_data no_entries_msg
-    elsif whois_record.body.blank?
-      logger.info "#{ip}: requested: #{data} [searched by: #{name}; Record found with id: #{whois_record.try(:id)} but body was EMPTY]"
-      send_data no_body_msg
-    else
+    if whois_record
       logger.info "#{ip}: requested: #{data} [searched by: #{name}; Record found with id: #{whois_record.try(:id)}]"
       send_data whois_record.unix_body
+    else
+      logger.info "#{ip}: requested: #{data} [searched by: #{name}; No record found]"
+      send_data no_entries_msg
     end
     close_connection_after_writing
   end
